+++ src/AbstractCommand.h
@@ -87,7 +87,7 @@ private:
   bool incNumConnection_;
   Timer serverStatTimer_;
 
-  size_t calculateMinSplitSize() const;
+  int32_t calculateMinSplitSize() const;
   void useFasterRequest(const SharedHandle<Request>& fasterRequest);
 #ifdef ENABLE_ASYNC_DNS
   void setNameResolverCheck(const SharedHandle<AsyncNameResolver>& resolver);
