+++ src/DefaultPieceStorage.h
@@ -234,7 +234,7 @@ public:
 
   virtual SharedHandle<DiskAdaptor> getDiskAdaptor();
 
-  virtual size_t getPieceLength(size_t index);
+  virtual int32_t getPieceLength(size_t index);
 
   virtual void advertisePiece(cuid_t cuid, size_t index);
