+++ src/GrowSegment.h
@@ -42,7 +42,7 @@ namespace aria2 {
 class GrowSegment:public Segment {
 private:
   SharedHandle<Piece> piece_;
-  size_t writtenLength_;
+  int32_t writtenLength_;
 public:
   GrowSegment(const SharedHandle<Piece>& piece);
 
@@ -68,27 +68,29 @@ public:
     return writtenLength_;
   }
 
-  virtual size_t getLength() const
+  virtual int32_t getLength() const
   {
     return 0;
   }
 
-  virtual size_t getSegmentLength() const
+  virtual int32_t getSegmentLength() const
   {
     return 0;
   }
 
-  virtual size_t getWrittenLength() const
+  virtual int32_t getWrittenLength() const
   {
     return writtenLength_;
   }
 
-  virtual void updateWrittenLength(size_t bytes);
+  virtual void updateWrittenLength(int32_t bytes);
 
 #ifdef ENABLE_MESSAGE_DIGEST
 
-  virtual bool updateHash(uint32_t begin,
-                          const unsigned char* data, size_t dataLength)
+  virtual bool updateHash
+  (int32_t begin,
+   const unsigned char* data,
+   size_t dataLength)
   {
     return false;
   }
